+++ src/core/position_sensorfusion/position_sensorfusion.hpp
@@ -15,7 +15,7 @@ namespace position_sensorfusion
         std::uint16_t confidence_position = 0U;
         std::uint16_t confidence_heading = 0U;
 
-        std::uint8_t pose_id = 0U;
+        std::uint16_t pose_id = 0U;
         std::uint8_t branch_id = 0U;
     };
